Small cleanup: move Composite normalizer outside core, removes the need for an ignored_normalizers...
